+++ Numbers/NumberExtensions.cs
@@ -2,7 +2,7 @@
 
 public static class NumberExtensions
 {
-    public static bool IsDivisibleBy(this int number, int divisor) => number % divisor == 0;
+    public static bool IsDivisibleBy(this long number, long divisor) => number % divisor == 0;
 
-    public static bool IsEven(int number) => number.IsDivisibleBy(2);
+    public static bool IsEven(long number) => number.IsDivisibleBy(2);
 }
