+++ Problems/Problem0002.cs
@@ -4,11 +4,11 @@ namespace Problems;
 
 public static class Problem0002
 {
-    public static int Example() => GetSumOfEvenFibonacciNumbersUpTo(8);
+    public static long Example() => GetSumOfEvenFibonacciNumbersUpTo(8);
 
-    public static int Solution() => GetSumOfEvenFibonacciNumbersUpTo(4_000_000);
+    public static long Solution() => GetSumOfEvenFibonacciNumbersUpTo(4_000_000);
 
-    private static int GetSumOfEvenFibonacciNumbersUpTo(int threshold)
+    private static long GetSumOfEvenFibonacciNumbersUpTo(int threshold)
     {
         return Fibonacci.GetAllLessOrEqual(threshold)
             .Where(NumberExtensions.IsEven)
